backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg[binary]
python-dotenv
pydantic
python-jose[cryptography]
httpx
